requirements.txt
llama-cpp-python==0.1.35
huggingface-hub
streamlit==1.24.0
